# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PAP - Projeto Aliança Panorama API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Node {
  /** Hierarchical code e.g. "1", "11", "12", "111" */
  code: string;
  title: string;
  /** @nullable */
  parentCode: string | null;
  childCount: number;
  /** Depth level in the tree */
  level: number;
}

export interface NodeDetail {
  code: string;
  title: string;
  /** @nullable */
  subtitle: string | null;
  /** @nullable */
  content: string | null;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  parentCode: string | null;
  children: Node[];
  level: number;
}

export interface Note {
  id: number;
  /** @nullable */
  nodeCode: string | null;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  /** @nullable */
  nodeCode?: string | null;
  content: string;
}

export interface UpdateNoteBody {
  content?: string;
}

export type AchievementType =
  (typeof AchievementType)[keyof typeof AchievementType];

export const AchievementType = {
  explored: "explored",
  read: "read",
  exercise: "exercise",
  approved: "approved",
} as const;

export interface Achievement {
  id: number;
  code: string;
  title: string;
  description: string;
  type: AchievementType;
  /** @nullable */
  nodeCode: string | null;
  /** @nullable */
  earnedAt: string | null;
  earned: boolean;
}

export interface Progress {
  /** Node codes that have been opened/explored */
  openedNodes: string[];
  /** Node codes whose content was fully read */
  readNodes: string[];
  achievements: Achievement[];
  totalNodes: number;
  explorationPercent: number;
}

export interface Summary {
  nodesExplored: number;
  nodesRead: number;
  achievementsEarned: number;
  totalAchievements: number;
  totalNodes: number;
  explorationPercent: number;
  recentlyOpened: Node[];
}

export type ListNodesParams = {
  /**
   * Parent node code to list children of
   */
  parentCode?: string;
};

export type ListNotesParams = {
  /**
   * Filter notes by node code/subject
   */
  nodeCode?: string;
};