# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PAP - Projeto Aliança Panorama API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List root nodes or nodes by parent
 */
export const ListNodesQueryParams = zod.object({
  parentCode: zod.coerce
    .string()
    .optional()
    .describe("Parent node code to list children of"),
});

export const ListNodesResponseItem = zod.object({
  code: zod
    .string()
    .describe('Hierarchical code e.g. \"1\", \"11\", \"12\", \"111\"'),
  title: zod.string(),
  parentCode: zod.string().nullable(),
  childCount: zod.number(),
  level: zod.number().describe("Depth level in the tree"),
});
export const ListNodesResponse = zod.array(ListNodesResponseItem);

/**
 * @summary Get a knowledge node by code
 */
export const GetNodeParams = zod.object({
  code: zod.coerce.string(),
});

export const GetNodeResponse = zod.object({
  code: zod.string(),
  title: zod.string(),
  subtitle: zod.string().nullable(),
  content: zod.string().nullable(),
  imageUrl: zod.string().nullable(),
  parentCode: zod.string().nullable(),
  children: zod.array(
    zod.object({
      code: zod
        .string()
        .describe('Hierarchical code e.g. \"1\", \"11\", \"12\", \"111\"'),
      title: zod.string(),
      parentCode: zod.string().nullable(),
      childCount: zod.number(),
      level: zod.number().describe("Depth level in the tree"),
    }),
  ),
  level: zod.number(),
});

/**
 * @summary Get user exploration summary and stats
 */
export const GetSummaryResponse = zod.object({
  nodesExplored: zod.number(),
  nodesRead: zod.number(),
  achievementsEarned: zod.number(),
  totalAchievements: zod.number(),
  totalNodes: zod.number(),
  explorationPercent: zod.number(),
  recentlyOpened: zod.array(
    zod.object({
      code: zod
        .string()
        .describe('Hierarchical code e.g. \"1\", \"11\", \"12\", \"111\"'),
      title: zod.string(),
      parentCode: zod.string().nullable(),
      childCount: zod.number(),
      level: zod.number().describe("Depth level in the tree"),
    }),
  ),
});

/**
 * @summary List all notes
 */
export const ListNotesQueryParams = zod.object({
  nodeCode: zod.coerce
    .string()
    .optional()
    .describe("Filter notes by node code\/subject"),
});

export const ListNotesResponseItem = zod.object({
  id: zod.number(),
  nodeCode: zod.string().nullable(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListNotesResponse = zod.array(ListNotesResponseItem);

/**
 * @summary Create a new note
 */
export const CreateNoteBody = zod.object({
  nodeCode: zod.string().nullish(),
  content: zod.string(),
});

/**
 * @summary Update a note
 */
export const UpdateNoteParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateNoteBody = zod.object({
  content: zod.string().optional(),
});

export const UpdateNoteResponse = zod.object({
  id: zod.number(),
  nodeCode: zod.string().nullable(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a note
 */
export const DeleteNoteParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get user progress including opened nodes and achievements
 */
export const GetProgressResponse = zod.object({
  openedNodes: zod
    .array(zod.string())
    .describe("Node codes that have been opened\/explored"),
  readNodes: zod
    .array(zod.string())
    .describe("Node codes whose content was fully read"),
  achievements: zod.array(
    zod.object({
      id: zod.number(),
      code: zod.string(),
      title: zod.string(),
      description: zod.string(),
      type: zod.enum(["explored", "read", "exercise", "approved"]),
      nodeCode: zod.string().nullable(),
      earnedAt: zod.coerce.date().nullable(),
      earned: zod.boolean(),
    }),
  ),
  totalNodes: zod.number(),
  explorationPercent: zod.number(),
});

/**
 * @summary Mark a node as opened (explored)
 */
export const OpenNodeParams = zod.object({
  code: zod.coerce.string(),
});

export const OpenNodeResponse = zod.object({
  openedNodes: zod
    .array(zod.string())
    .describe("Node codes that have been opened\/explored"),
  readNodes: zod
    .array(zod.string())
    .describe("Node codes whose content was fully read"),
  achievements: zod.array(
    zod.object({
      id: zod.number(),
      code: zod.string(),
      title: zod.string(),
      description: zod.string(),
      type: zod.enum(["explored", "read", "exercise", "approved"]),
      nodeCode: zod.string().nullable(),
      earnedAt: zod.coerce.date().nullable(),
      earned: zod.boolean(),
    }),
  ),
  totalNodes: zod.number(),
  explorationPercent: zod.number(),
});

/**
 * @summary Mark a node content as read
 */
export const ReadNodeParams = zod.object({
  code: zod.coerce.string(),
});

export const ReadNodeResponse = zod.object({
  openedNodes: zod
    .array(zod.string())
    .describe("Node codes that have been opened\/explored"),
  readNodes: zod
    .array(zod.string())
    .describe("Node codes whose content was fully read"),
  achievements: zod.array(
    zod.object({
      id: zod.number(),
      code: zod.string(),
      title: zod.string(),
      description: zod.string(),
      type: zod.enum(["explored", "read", "exercise", "approved"]),
      nodeCode: zod.string().nullable(),
      earnedAt: zod.coerce.date().nullable(),
      earned: zod.boolean(),
    }),
  ),
  totalNodes: zod.number(),
  explorationPercent: zod.number(),
});

/**
 * @summary List all achievements/badges
 */
export const ListAchievementsResponseItem = zod.object({
  id: zod.number(),
  code: zod.string(),
  title: zod.string(),
  description: zod.string(),
  type: zod.enum(["explored", "read", "exercise", "approved"]),
  nodeCode: zod.string().nullable(),
  earnedAt: zod.coerce.date().nullable(),
  earned: zod.boolean(),
});
export const ListAchievementsResponse = zod.array(ListAchievementsResponseItem);